# book-network-api/src/main/java/pl/tkaczyk/booknetwork/book/BorrowedBookResponse.java
package pl.tkaczyk.booknetwork.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {

    private Long id;
    private String title;
    private String authorName;
    private String isbn;
    private double rate;
    private boolean returned;
    private boolean returnApproved;
}