# book-network-api/src/main/java/pl/tkaczyk/booknetwork/feedback/FeedbackResponse.java
package pl.tkaczyk.booknetwork.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}